Change ordering of indices in radii and filling vectors
In the past, the first radius in the radii vector was the outermost
radius (likewise the filling with the expection that the embedding is
stored in an extra vector).

This commit changes to order to match the ordering of Jin and what
we use in the documentation.

TODO: one could think of adding the embedding to the filling vector.
We still have the interface embedding() and filling() so that nothing
changes for the user, but makes internal handling potentially easier.